Create a super type for Constructor- and MethodDeclaration - #738
